Refactor dbt model building functionality into ModelSplitButton
- Removed BuildDbtSplitButton component and integrated its functionality into ModelSplitButton for a more cohesive user experience.
- Added build options for upstream, downstream, and both directions directly within ModelSplitButton.
- Updated ProjectDetails screen to utilize the refactored ModelSplitButton, enhancing the interface for dbt model actions.
- Cleaned up imports and removed obsolete code related to the deleted BuildDbtSplitButton.
