+++ src/renderer/components/dbtModelButtons/ModelSplitButton.tsx
@@ -51,6 +51,7 @@ export const ModelSplitButton: React.FC<ModelSplitButtonProps> = ({
     test: dbtTestModel,
     isRunning: isRunningDbtModel,
     list: dbtList,
+    build: dbtBuildModel,
   } = useDbt();
 
   const handleCompileModel = async () => {
@@ -406,6 +407,112 @@ export const ModelSplitButton: React.FC<ModelSplitButtonProps> = ({
     }
   };
 
+  const handleBuildModel = async () => {
+    if (!isDbtConfigured) {
+      toast.info('Please configure dbt path in settings');
+      return;
+    }
+
+    try {
+      // Extract model name from path for single model building
+      const modelName = extractModelNameFromPath(modelPath);
+      if (!modelName) {
+        toast.error('Could not extract model name from path');
+        return;
+      }
+
+      // Build the single model using dbt build --select
+      // This will run the model + tests + seeds + snapshots
+      await dbtBuildModel(project, modelName);
+      toast.success(`Model '${modelName}' built successfully with tests`);
+    } catch (error) {
+      const errorMessage =
+        error instanceof Error ? error.message : 'Unknown error';
+      toast.error(`Model build failed: ${errorMessage}`);
+    }
+  };
+
+  const handleBuildModelDownstream = async () => {
+    if (!isDbtConfigured) {
+      toast.info('Please configure dbt path in settings');
+      return;
+    }
+
+    try {
+      // Extract model name from path for downstream building
+      const modelName = extractModelNameFromPath(modelPath);
+      if (!modelName) {
+        toast.error('Could not extract model name from path');
+        return;
+      }
+
+      // Build the model and all its downstream dependencies using dbt build --select model_name+
+      // The + suffix tells dbt to include all downstream models
+      await dbtBuildModel(project, `${modelName}+`);
+      toast.success(
+        `Model '${modelName}' and downstream models built successfully`,
+      );
+    } catch (error) {
+      const errorMessage =
+        error instanceof Error ? error.message : 'Unknown error';
+      toast.error(`Downstream build failed: ${errorMessage}`);
+    }
+  };
+
+  const handleBuildModelUpstream = async () => {
+    if (!isDbtConfigured) {
+      toast.info('Please configure dbt path in settings');
+      return;
+    }
+
+    try {
+      // Extract model name from path for upstream building
+      const modelName = extractModelNameFromPath(modelPath);
+      if (!modelName) {
+        toast.error('Could not extract model name from path');
+        return;
+      }
+
+      // Build the model and all its upstream dependencies using dbt build --select +model_name
+      // The + prefix tells dbt to include all upstream models (parents)
+      await dbtBuildModel(project, `+${modelName}`);
+      toast.success(
+        `Model '${modelName}' and upstream models built successfully`,
+      );
+    } catch (error) {
+      const errorMessage =
+        error instanceof Error ? error.message : 'Unknown error';
+      toast.error(`Upstream build failed: ${errorMessage}`);
+    }
+  };
+
+  const handleBuildModelBothDirections = async () => {
+    if (!isDbtConfigured) {
+      toast.info('Please configure dbt path in settings');
+      return;
+    }
+
+    try {
+      // Extract model name from path for both directions building
+      const modelName = extractModelNameFromPath(modelPath);
+      if (!modelName) {
+        toast.error('Could not extract model name from path');
+        return;
+      }
+
+      // Build the model and all its upstream and downstream dependencies using dbt build --select +model_name+
+      // The + prefix and suffix tells dbt to include both upstream and downstream models
+      await dbtBuildModel(project, `+${modelName}+`);
+      toast.success(
+        `Model '${modelName}' and all related models (upstream + downstream) built successfully`,
+      );
+    } catch (error) {
+      const errorMessage =
+        error instanceof Error ? error.message : 'Unknown error';
+      toast.error(`Full dependency build failed: ${errorMessage}`);
+    }
+  };
+
   return (
     <>
       <SplitButton
@@ -429,6 +536,18 @@ export const ModelSplitButton: React.FC<ModelSplitButtonProps> = ({
         }
         leftIcon={<DirectionsRun />}
         menuItems={[
+          {
+            name: 'Compile',
+            onClick: handleCompileModel,
+            leftIcon: <Icon src={icons.dbtTm} width={16} height={16} />,
+            subTitle: 'Compile the dbt model',
+          },
+          {
+            name: 'Preview',
+            onClick: handlePreviewModel,
+            leftIcon: <Icon src={icons.dbtTm} width={16} height={16} />,
+            subTitle: 'Preview the dbt model data',
+          },
           {
             name: 'Run',
             onClick: handleRunModel,
@@ -455,6 +574,31 @@ export const ModelSplitButton: React.FC<ModelSplitButtonProps> = ({
               'Run the model and all its upstream and downstream dependencies',
           },
           {
+            name: 'Build Model',
+            onClick: handleBuildModel,
+            leftIcon: <Icon src={icons.dbtTm} width={16} height={16} />,
+            subTitle: 'Build model with tests and validation',
+          },
+          {
+            name: 'Build model+ (Downstream)',
+            onClick: handleBuildModelDownstream,
+            leftIcon: <Icon src={icons.dbtTm} width={16} height={16} />,
+            subTitle: 'Build the model and all its downstream dependencies',
+          },
+          {
+            name: 'Build +model (Upstream)',
+            onClick: handleBuildModelUpstream,
+            leftIcon: <Icon src={icons.dbtTm} width={16} height={16} />,
+            subTitle: 'Build the model and all its upstream dependencies',
+          },
+          {
+            name: 'Build +model+ (Up/downstream)',
+            onClick: handleBuildModelBothDirections,
+            leftIcon: <Icon src={icons.dbtTm} width={16} height={16} />,
+            subTitle:
+              'Build the model and all its upstream and downstream dependencies',
+          },
+          {
             name: 'Test',
             onClick: handleTestModel,
             leftIcon: <Icon src={icons.dbtTm} width={16} height={16} />,
@@ -478,18 +622,6 @@ export const ModelSplitButton: React.FC<ModelSplitButtonProps> = ({
             leftIcon: <Icon src={icons.dbtTm} width={16} height={16} />,
             subTitle:
               'Test the model and all its upstream and downstream dependencies',
-          },
-          {
-            name: 'Compile',
-            onClick: handleCompileModel,
-            leftIcon: <Icon src={icons.dbtTm} width={16} height={16} />,
-            subTitle: 'Compile the dbt model',
-          },
-          {
-            name: 'Preview',
-            onClick: handlePreviewModel,
-            leftIcon: <Icon src={icons.dbtTm} width={16} height={16} />,
-            subTitle: 'Preview the dbt model data',
           },
         ]}
       />
