+++ src/renderer/components/dbtModelButtons/index.ts
@@ -1,3 +1,2 @@
 export { ModelSplitButton } from './ModelSplitButton';
-export { BuildDbtSplitButton } from './BuildDbtSplitButton';
 export { ProjectDbtSplitButton } from './ProjectDbtSplitButton';
